Move ui_graphics_quad to shape_renderer
